orchestrator/requirements.txt
# Minimal orchestrator dependencies
langchain>=0.3.10
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-anthropic>=0.1.0

# Provider SDKs
openai>=1.54.0
anthropic>=0.34.0
google-generativeai>=0.8.0

# Utilities
PyYAML>=5.3
requests>=2.32.0
python-dotenv>=1.0.0

# TUI support
rich>=13.0.0

# Local model helpers (optional - only if using Ollama workers/orchestrator)
ollama>=0.4.0
langchain-ollama>=0.1.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0
pytest-mock>=3.12.0
